🚚 [projects] Move `ProjectConfig` from `templates` to `projects`
