+++ tests/functional/test_link.py
@@ -3,9 +3,9 @@ from pathlib import Path
 
 import pytest
 
+from cutty.projects.projectconfig import readprojectconfigfile
 from cutty.projects.repository import LATEST_BRANCH
 from cutty.projects.repository import UPDATE_BRANCH
-from cutty.templates.adapters.cookiecutter.projectconfig import readprojectconfigfile
 from cutty.util.git import Repository
 from tests.functional.conftest import RunCutty
 from tests.functional.conftest import RunCuttyError
